Add ignite configuration file support to ignited
Adds a new package `pkg/config` to have common configuration handling
code.
